+++ src/olmo_core/launch/beaker.py
@@ -96,6 +96,16 @@ class OLMoCoreBeakerImage(StrEnum):
     The stable image with CUDA pinned to 12.8.
     """
 
+    tch280_cu128 = "olmo-core-tch280cu128-2025-09-18"
+    """
+    Built with torch 2.8.0 and CUDA 12.8.
+    """
+
+    tch280_cu129 = "olmo-core-tch280cu129-2025-09-18"
+    """
+    Built with torch 2.8.0 and CUDA 12.9.
+    """
+
 
 @dataclass
 class BeakerEnvVar(Config):